+++ src/main/java/com/snocal/sms/conf/core/repository/MenuRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.snocal.sms.conf.core.entity.MenuEntity;
 
 @Repository
-public interface MenuRepository extends JpaRepository<MenuEntity, Integer> {
+public interface MenuRepository extends JpaRepository<MenuEntity, Long> {
 
 }
